requirements.txt
aiocache==0.12.2
aiofiles==23.2.1
aiohttp==3.8.5
aiosqlite==0.19.0
captcha==0.5.0
chat-exporter==2.6.1
colorama==0.4.4
colorlog==6.7.0
easy-pil==0.3.0
function-cooldowns==2.0.1
geopy==2.4.0
humanfriendly==10.0
mafic==2.9.3
orjson==3.9.7
Pillow==9.5.0
PyNaCl==1.5.0
pyright
git+https://github.com/xXenvy/async-scrapetube
git+https://github.com/xXenvy/cordcutter.git
git+https://github.com/SmiffyBot/nextcord.git@Smiffy
